employee/src/main/java/employee/domain/model/Client.java
package employee.domain.model;

import java.util.List;

public class Client {

    private String name;
    private List<Product> products;
    private Integer productLevel;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public List<Product> getProducts() {
        return products;
    }

    public void setProducts(List<Product> products) {
        this.products = products;
    }

    public Integer getProductLevel() {
        return productLevel;
    }

    public void setProductLevel(Integer productLevel) {
        this.productLevel = productLevel;
    }

    @Override
    public String toString() {
        return "Client{" +
                "name='" + name + '\'' +
                ", products=" + products +
                ", productLevel=" + productLevel +
                '}';
    }
}
